### 2025-05-02 #8149 @editorialbot
[PRE REVIEW]: geospaNN: a Python package for geospatial neural networks
**Submitting author:** @WentaoZhan1998 (<a href="http://orcid.org/0009-0002-1183-5305">Wentao Zhan</a>)
**Repository:** https://github.com/WentaoZhan1998/geospaNN/
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.9
