### 2025-05-09 127c6f9
Reduced the words for joss
