+++ joss/paper.md
@@ -30,41 +30,43 @@ Y = X\beta + \epsilon(s)
 $$
 has long been the standard approach to model such data. SPLMM accounts for the fixed effects between observations $Y$ and covariates $X$ via the linear regression part, while accounting for the correlation across spatial locations $s$ via the spatial error process $\epsilon(s)$. Typically, the spatial dependency embedded in $\epsilon(s)$ is modeled using a Gaussian Process, which provides a parsimonious and flexible solution to modeling spatial correlations and provide spatial predictions at new locations via kriging. 
 
-The linear regression part of SPLMM is restrictive. Recently, machine learning techniques like artificial neural networks have been increasingly incorporated into geostatistics to address complex and non-linear interactions among variables. However, few methods were originally designed for dependent data structures and cannot directly encode correlation. Even those modified to account for spatial dependencies, like adding a large number of spatial features as inputs to neural network, still face scalability limitations due to the expanded dimensionality from adding many features. To address these concerns, we introduce `geospaNN`, a Python software for analysis of geospatial data using neural networks. `geospaNN` implements NN-GLS, a novel adaptation of Neural Networks (NN) proposed in @zhan2024neural. NN-GLS bridges ideas of geostatistical modeling and machine learning by embedding neural network with spatial mixed models, retaining the capacity to directly model spatial correlations. NN-GLS extends the SPLMM to non-linear scenarios by replacing $X\beta$ with $m(X)$, a non-linear function. NN-GLS simultaneously performs non-linear mean function estimation using a neural network and spatial predictions (including prediction intervals) using Gaussian processes. For computational efficiency, `geospaNN` leverages the widely popular Nearest Neighbor Gaussian Process (NNGP) approximation [@datta2016nearest, @finley2019efficient]. It is also, to our knowledge, the first Python package to implement NNGP for scalable covariance matrix computations in geostatistical mmodels. This is a standalone component of broad and  independent utility. <!--benefiting other geospatial computation tools. -->
+Recently, machine learning techniques like Neural Networks (NNs) have been increasingly incorporated into geostatistics to address complex and non-linear interactions among variables. In this paper, we introduce `geospaNN`, a Python software for analysis of geospatial data. `geospaNN` implements a novel adaptation of NN <!--proposed in @zhan2024neural that extends the SPLMM to non-linear scenarios while retaining the capacity to directly model spatial correlations. `geospaNN`--> and simultaneously performs non-linear mean function estimation using a NN and spatial predictions (including prediction intervals) using Gaussian processes. For computational efficiency, `geospaNN` leverages the widely popular Nearest Neighbor Gaussian Process (NNGP) approximation [@datta2016nearest, @finley2019efficient]. It is also, to our knowledge, the first Python package to implement scalable covariance matrix computations in geostatistical models. <!--This is a standalone component of broad and  independent utility. benefiting other geospatial computation tools. -->
 
 
 
 ### Statement of Need
 
[…]
