### 2025-06-04 #8149 @kks32
@vissarion I'm happy to review.
