### 2025-10-26 #8389 @vissarion
:wave: @benyamindsmith, @mszell, @professornaite, @Exelegcho1 would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html. Please be so kind to let me know even if you're not interested, so that I can start looking for other reviewers.
