### 2025-10-31 #8389 @cbueth
Hello @vissarion, giving a quick look the package seems solid, but still quite lean on the testing side. Regarding the NN falling outside my fields I do not think I feel comfortable reviewing it.
