### 2025-11-04 #8389 @WentaoZhan1998
Hi @vissarion, shall we ask [@ArkajyotiSaha](https://github.com/ArkajyotiSaha)? I know he's an expert in spatial machine learning and should an ideal candidate.
