### 2025-11-08 #8389 @professornaite
I will not be able to review until later in spring. Thank you.

On Sun, Oct 26, 2025 at 3:24 AM Vissarion Fisikopoulos <
***@***.***> wrote:

> *vissarion* left a comment (openjournals/joss-reviews#8389)
> <https://github.com/openjournals/joss-reviews/issues/8389#issuecomment-3448086442>
>
> 👋 @benyamindsmith <https://github.com/benyamindsmith>, @mszell
> <https://github.com/mszell>, @professornaite
> <https://github.com/professornaite>, @Exelegcho1
> <https://github.com/Exelegcho1> would any of you be willing to review
> this submission for JOSS? We carry out our checklist-driven reviews here in
> GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html. Please be so
> kind to let me know even if you're not interested, so that I can start
> looking for other reviewers.
>
