+++ README.md
@@ -53,75 +53,74 @@ Nearest Neighbor Gaussian Process (NNGP) (Datta et al., 2016) which makes it sui
 </a>
 </div>
 
-## Temporary notes (Updated on Sep 12th 2024)
-1. The installation of the package relies on PyTorch and PyG libraries, which for now have to be installed manually.
+## Temporary notes (Updated on Nov 2025)
+1. The installation of the package relies is now based on PyTorch 2.7.0.
 
 ## Installation
-### Create and enter virtual environment (recommended)
-1. If you haven't installed anaconda on your machine, refer to this [doc](https://docs.anaconda.com/anaconda/install/), follow the instruction, 
+We provide two straightforward installation approaches: via conda and via pip.
+Depending on your system setup, it is possible to combine both methods, but be aware that mixing Conda and Pip installations can sometimes lead to dependency conflicts. Proceed with caution and ensure that package versions remain compatible.
+
+### Approach 1: all-in-one through conda (recommended)
+1. If you haven't installed anaconda on your machine, refer to this [doc](https://docs.anaconda.com/anaconda/install/) follow the instruction 
 and install the right version.
-2. Create the conda virtual environment. Refer to this [doc](https://conda.io/projects/conda/en/latest/user-guide/tasks/manage-environments.html). Example:
+2. Create the conda virtual environment from the [environment.yml](https://github.com/WentaoZhan1998/geospaNN/blob/main/environment.yml) file in this repository. You can specify your environment name by editing "env_name" on the first line of the yml file.
+Example:
 ```commandline\ 
 # bash
-conda create -n [name of your environment] python=3.10
+conda env create -f environment.yml
 ```
+Note:
+For Apple Silicon users on a Mac with an Apple M-series (ARM64) chip, you can improve performance by explicitly creating the environment for the ARM architecture instead:
+```commandline\ 
+# bash
+conda env create -f environment.yml --subdir osx-arm64
+```
+For more details on creating a conda environment, refer to this [doc](https://conda.io/projects/conda/en/latest/user-guide/tasks/manage-environments.html).
 3. Enter the virtual environment by running:
 ```commandline\ 
 # bash
 conda activate [name of your environment]
 ```
-4. In the current version of geospaNN, to use the R-package [BRISC](https://github.com/ArkajyotiSaha/BRISC) 
-for spatial parameter estimation (through rpy2), we need R installed in the environment. In order to install R, simply run:
-```commandline\ 
-# bash
-conda install r-base
-```
-If you already have native R installed, it's also possible to manually initialize R for rpy2. 
-See [here](https://rpy2.github.io/doc/latest/html/overview.html#install-installation) for more details.
 
-### Manual dependency installation
-(Currently) to install the development version of the package, a pre-installed PyTorch and PyG libraries are needed.
-We provide options to install PyG libraries using conda and pip.
+### Approach 2: using pip
+(Currently) to avoid running issue, matched PyTorch and PyG libraries are needed, requiring us to install torch and pyg library manually.
 
-#### Option 1: Using Conda
-For conda, installation in the following order is recommended. It may take around 10 minutes for conda to solve the environment for pytorch-sparse.
-The following chunk has been tested in a python 3.10 environment.
-```
-#bash
-conda install pytorch torchvision -c pytorch
-conda install pyg -c pyg        
-conda install pytorch-sparse -c pyg 
-```
-
-#### Option 2: Using pip
[…]
