+++ joss/paper.bib
@@ -211,5 +211,23 @@
   publisher={Wiley Online Library}
 }
 
+@inproceedings{murphy2014pykrige,
+  title={PyKrige: development of a kriging toolkit for Python},
+  author={Murphy, Benjamin S},
+  booktitle={AGU fall meeting abstracts},
+  volume={2014},
+  pages={H51K--0753},
+  year={2014}
+}
+
+@article{gardner2018gpytorch,
+  title={Gpytorch: Blackbox matrix-matrix gaussian process inference with gpu acceleration},
+  author={Gardner, Jacob and Pleiss, Geoff and Weinberger, Kilian Q and Bindel, David and Wilson, Andrew G},
+  journal={Advances in neural information processing systems},
+  volume={31},
+  year={2018}
+}
+
+
 
 
