+++ joss/paper.md
@@ -10,7 +10,7 @@ tags:
 output: pdf_document
 authors:
 - name: Wentao Zhan
-  affiliation: 1
+  affiliation: 2
 - name: Abhirup Datta
   affiliation: 1
 bibliography: paper.bib
@@ -20,34 +20,52 @@ editor_options:
 affiliations:
 - name: Department of Biostatistics, Johns Hopkins Bloomberg School of Public Health
   index: 1
+- name: Department of Statistics, University of Wisconsin-Madison
+  index: 2
 ---
 
 ### Summary
-
+<!--
 Geostatistical models are widely used to analyse datasets with spatial information encountered frequently in the geosciences (climate, ecology, forestry, environmental sciences, and other research fields). In geostatistics, the spatial linear mixed effects model (SPLMM)\
 $$
 Y = X\beta + \epsilon(s)
 $$
 has long been the standard approach to model such data. SPLMM accounts for the fixed effects between observations $Y$ and covariates $X$ via the linear regression part, while accounting for the correlation across spatial locations $s$ via the spatial error process $\epsilon(s)$. Typically, the spatial dependency embedded in $\epsilon(s)$ is modeled using a Gaussian Process, which provides a parsimonious and flexible solution to modeling spatial correlations and provide spatial predictions at new locations via kriging. 
 
-Recently, machine learning techniques like Neural Networks (NNs) have been increasingly incorporated into geostatistics to address complex and non-linear interactions among variables. In this paper, we introduce `geospaNN`, a Python software for analysis of geospatial data. `geospaNN` implements a novel adaptation of NN <!--proposed in @zhan2024neural that extends the SPLMM to non-linear scenarios while retaining the capacity to directly model spatial correlations. `geospaNN`--> and simultaneously performs non-linear mean function estimation using a NN and spatial predictions (including prediction intervals) using Gaussian processes. For computational efficiency, `geospaNN` leverages the widely popular Nearest Neighbor Gaussian Process (NNGP) approximation [@datta2016nearest, @finley2019efficient]. It is also, to our knowledge, the first Python package to implement scalable covariance matrix computations in geostatistical models. <!--This is a standalone component of broad and  independent utility. benefiting other geospatial computation tools. -->
+Recently, machine learning techniques like Neural Networks (NNs) have been increasingly incorporated into geostatistics to address complex and non-linear interactions among variables. In this paper, we introduce `geospaNN`, a Python software for analysis of geospatial data. `geospaNN` implements a novel adaptation of NN <!--proposed in @zhan2024neural that extends the SPLMM to non-linear scenarios while retaining the capacity to directly model spatial correlations. `geospaNN`--> <!--and simultaneously performs non-linear mean function estimation using a NN and spatial predictions (including prediction intervals) using Gaussian processes. For computational efficiency, `geospaNN` leverages the widely popular Nearest Neighbor Gaussian Process (NNGP) approximation [@datta2016nearest, @finley2019efficient]. It is also, to our knowledge, the first Python package to implement scalable covariance matrix computations in geostatistical models. <!--This is a standalone component of broad and  independent utility. benefiting other geospatial computation tools. -->
 
-
[…]
