+++ joss/paper.bib
@@ -104,9 +104,10 @@
   title={Neural networks for geospatial data},
   author={Zhan, Wentao and Datta, Abhirup},
   journal={Journal of the American Statistical Association},
-  number={just-accepted},
-  pages={1--21},
-  year={2024},
+  volume={120},
+  number={549},
+  pages={535--547},
+  year={2025},
   publisher={Taylor \& Francis}
 }
 
