### 2025-12-04 #8389 @WentaoZhan1998
@vissarion @ArkajyotiSaha @Exelegcho1 We thank the editor and the reviewers for their positive feedback which have helped us to improve the software and manuscript. We tried to address all the reviewer comments in the paper. Updated versions of the package and the paper are now available in the GitHub repository (https://github.com/WentaoZhan1998/geospaNN).
We revised the manuscript thoroughly according to the feedback from @ArkajyotiSaha. A formal response letter is available in https://github.com/WentaoZhan1998/geospaNN/blob/main/joss/JOSS_author_response.pdf.
Please feel free to let me know if you have any questions or further comments. Thanks for your time and consideration!
