+++ joss/paper.bib
@@ -6,7 +6,8 @@
   number={5},
   pages={162--171},
   year={2016},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/wics.1383}
 }
 
 @article{datta2022nearest,
@@ -17,7 +18,8 @@
   number={5},
   pages={e1574},
   year={2022},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/wics.1574}
 }
 
 @article{finley2019efficient,
@@ -28,7 +30,8 @@
   number={2},
   pages={401--414},
   year={2019},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi={10.1080/10618600.2018.1537924}
 }
 
 @article{finley2022spnngp,
@@ -48,7 +51,8 @@
   number={3},
   pages={ujae089},
   year={2024},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/biomtc/ujae089}
 }
 
 @article{georganos2021geographical,
@@ -59,7 +63,8 @@
   number={2},
   pages={121--136},
   year={2021},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi={10.1080/10106049.2019.1595177}
 }
 
 @article{heaton2024adjusting,
@@ -77,7 +82,8 @@
   number={12},
   pages={e0315127},
   year={2024},
-  publisher={Public Library of Science San Francisco, CA USA}
+  publisher={Public Library of Science San Francisco, CA USA},
+  doi={10.31223/x53m6t}
 }
 
 @article{paszke2019pytorch,
@@ -96,7 +102,8 @@
   number={1},
   pages={e184},
   year={2018},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/sta4.184}
 }
 
 
@@ -108,14 +115,16 @@
   number={549},
   pages={535--547},
   year={2025},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi={10.1080/01621459.2024.2356293}
 }
 
 @article{chen2020deepkriging,
   title={Deepkriging: Spatially dependent deep neural networks for spatial prediction},
   author={Chen, Wanfang and Li, Yuxiao and Reich, Brian J and Sun, Ying},
   journal={arXiv preprint arXiv:2007.11972},
-  year={2020}
+  year={2020},
+  doi={10.5705/ss.202021.0277}
 }
 
 @inproceedings{fan2022gnn,
@@ -125,7 +134,8 @@
   volume={36},
   number={11},
   pages={11873--11881},
-  year={2022}
+  year={2022},
+  doi={10.1609/aaai.v36i11.21444}
 }
 
 @article{stein2004approximating,
@@ -136,14 +146,16 @@
   number={2},
   pages={275--296},
   year={2004},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1046/j.1369-7412.2003.05512.x}
 }
 
 @article{tonks2022forecasting,
   title={Forecasting West Nile Virus with Graph Neural Networks: Harnessing Spatial Dependence in Irregularly Sampled Geospatial Data},
   author={Tonks, Adam and Harris, Trevor and Li, Bo and Brown, William and Smith, Rebecca},
   journal={arXiv preprint arXiv:2212.11367},
-  year={2022}
+  year={2022},
+  doi={10.1029/2023gh000784}
 }
 
 @inproceedings{wang2020traffic,
@@ -151,7 +163,8 @@
[…]
