+++ joss/paper.bib
@@ -10,16 +10,22 @@
   doi={10.1002/wics.1383}
 }
 
-@article{datta2022nearest,
-  title={Nearest-neighbor sparse {C}holesky matrices in spatial statistics},
-  author={Datta, Abhirup},
-  journal={Wiley Interdisciplinary Reviews: Computational Statistics},
-  volume={14},
-  number={5},
-  pages={e1574},
+@inproceedings{fan2022gnn,
+  title={A GNN-RNN approach for harnessing geospatial and temporal information: application to crop yield prediction},
+  author={Fan, Joshua and Bai, Junwen and Li, Zhiyun and Ortiz-Bobea, Ariel and Gomes, Carla P},
+  booktitle={Proceedings of the AAAI conference on artificial intelligence},
+  volume={36},
+  number={11},
+  pages={11873--11881},
   year={2022},
-  publisher={Wiley Online Library},
-  doi={10.1002/wics.1574}
+  doi={10.1609/aaai.v36i11.21444}
+}
+
+@article{PyG2019,
+  author    = {Matthias Fey and Jan E. Lenssen},
+  title     = {Fast Graph Representation Learning with {PyTorch Geometric}},
+  journal   = {arXiv preprint arXiv:1903.02428},
+  year      = {2019},
 }
 
 @article{finley2019efficient,
@@ -34,44 +40,13 @@
   doi={10.1080/10618600.2018.1537924}
 }
 
-@article{finley2022spnngp,
-  title={spNNGP R package for nearest neighbor Gaussian process models},
-  author={Finley, Andrew O and Datta, Abhirup and Banerjee, Sudipto},
-  journal={Journal of Statistical Software},
-  volume={103},
-  pages={1--40},
-  year={2022}
-}
-
-@article{gilbert2024visibility,
-  title={Visibility graph-based covariance functions for scalable spatial analysis in non-convex partially Euclidean domains},
-  author={Gilbert, Brian and Datta, Abhirup},
-  journal={Biometrics},
-  volume={80},
-  number={3},
-  pages={ujae089},
-  year={2024},
-  publisher={Oxford University Press},
-  doi={10.1093/biomtc/ujae089}
-}
-
-@article{georganos2021geographical,
-  title={Geographical random forests: a spatial extension of the random forest algorithm to address spatial heterogeneity in remote sensing and population modelling},
-  author={Georganos, Stefanos and Grippa, Tais and Niang Gadiaga, Assane and Linard, Catherine and Lennert, Moritz and Vanhuysse, Sabine and Mboga, Nicholus and Wolff, El{\'e}onore and Kalogirou, Stamatis},
-  journal={Geocarto International},
-  volume={36},
-  number={2},
-  pages={121--136},
-  year={2021},
-  publisher={Taylor \& Francis},
-  doi={10.1080/10106049.2019.1595177}
-}
-
-@article{heaton2024adjusting,
-  title={Adjusting for Spatial Correlation in Machine and Deep Learning},
-  author={Heaton, Matthew J and Millane, Andrew and Rhodes, Jake S},
-  journal={arXiv preprint arXiv:2410.04312},
-  year={2024}
+@article{gardner2018gpytorch,
+  title={Gpytorch: Blackbox matrix-matrix gaussian process inference with gpu acceleration},
+  author={Gardner, Jacob and Pleiss, Geoff and Weinberger, Kilian Q and Bindel, David and Wilson, Andrew G},
+  journal={Advances in neural information processing systems},
+  volume={31},
+  year={2018},
+  doi={10.5555/3327757.3327857}
 }
 
 @article{maxwell2024geodl,
@@ -84,6 +59,15 @@
[…]
