+++ joss/paper.bib
@@ -45,8 +45,7 @@
   author={Gardner, Jacob and Pleiss, Geoff and Weinberger, Kilian Q and Bindel, David and Wilson, Andrew G},
   journal={Advances in neural information processing systems},
   volume={31},
-  year={2018},
-  doi={10.5555/3327757.3327857}
+  year={2018}
 }
 
 @article{maxwell2024geodl,
