### 2026-01-28 #8389 @crvernon
@WentaoZhan1998 - I just need you to address the following before I accept:

In your paper:

- [x] LINE 105:  The code block runs off the page once the PDF is compiled.
- [x] LINE 158:  The "b" in "bayesian" and the "g" in "gaussian" should be capitalized.  You can do this in your bib file by placing curly brackets around the characters you wish to maintain the formatting of.
- [x] LINE 162:  The "g" in "gaussian" should be capitalized also "gpu" should be capitalized as "GPU"
- [x] LINE 164:  The "r" as in the R programming language should be capitalized.
- [x] LINE 167:  The "p" in "python" should be capitalized

Let me know when these changes have been made and after I confirm I will accept.
