+++ joss/paper.bib
@@ -29,7 +29,7 @@
 }
 
 @article{finley2019efficient,
-  title={Efficient algorithms for Bayesian nearest neighbor Gaussian processes},
+  title={Efficient algorithms for {B}ayesian nearest neighbor {G}aussian processes},
   author={Finley, Andrew O and Datta, Abhirup and Cook, Bruce D and Morton, Douglas C and Andersen, Hans E and Banerjee, Sudipto},
   journal={Journal of Computational and Graphical Statistics},
   volume={28},
@@ -41,7 +41,7 @@
 }
 
 @article{gardner2018gpytorch,
-  title={Gpytorch: Blackbox matrix-matrix gaussian process inference with gpu acceleration},
+  title={GPyTorch: {B}lackbox matrix-matrix {G}aussian process inference with {GPU} acceleration},
   author={Gardner, Jacob and Pleiss, Geoff and Weinberger, Kilian Q and Bindel, David and Wilson, Andrew G},
   journal={Advances in neural information processing systems},
   volume={31},
@@ -49,7 +49,7 @@
 }
 
 @article{maxwell2024geodl,
-  title={geodl: An R package for geospatial deep learning semantic segmentation using torch and terra},
+  title={geodl: An {R} package for geospatial deep learning semantic segmentation using torch and terra},
   author={Maxwell, Aaron E and Farhadpour, Sarah and Das, Srinjoy and Yang, Yalin},
   journal={PloS one},
   volume={19},
@@ -61,7 +61,7 @@
 }
 
 @inproceedings{murphy2014pykrige,
-  title={PyKrige: development of a kriging toolkit for Python},
+  title={PyKrige: development of a kriging toolkit for {P}ython},
   author={Murphy, Benjamin S},
   booktitle={AGU fall meeting abstracts},
   volume={2014},
