+++ geospaNN/utils.py
@@ -102,14 +102,13 @@ class LRScheduler():
                     min_lr=self.min_lr
                 )
         else:
-            self.lr_scheduler = torch.optim.lr_scheduler.CosineAnnealingLR(
-                self.optimizer,
-                mode='min',
-                patience=self.patience,
-                factor=self.factor,
-                min_lr=self.min_lr,
-                verbose=True
-            )
+            self.lr_scheduler = torch.optim.lr_scheduler.ReduceLROnPlateau(
+                    self.optimizer,
+                    mode='min',
+                    patience=self.patience,
+                    factor=self.factor,
+                    min_lr=self.min_lr
+                )
 
     def __call__(self, val_loss):
         self.lr_scheduler.step(val_loss)
