### 2026-01-29 #8389 @crvernon
@WentaoZhan1998 please remove the special formatting (e.g., strong/bold) from the title of your paper.  It causes a formatting issue when we publish.  Let me know when this is done and I'll reaccept.
